refactor: move TextEditor component
