+++ src/app/_components/AppShell.tsx
@@ -8,8 +8,6 @@ import { useDisclosure } from "@mantine/hooks";
 import styles from "./AppShell.module.css";
 import Sidebar from "./Sidebar";
 
-import TextEditor from "#/app/channels/[channel_id]/_components/TextEditor";
-
 type AppShellProps = {
   children: React.ReactNode;
 };
@@ -20,8 +18,7 @@ const AppShell: React.FC<AppShellProps> = ({ children }) => {
   return (
     <MantineAppShell
       layout="alt"
-      header={{ height: 60 }}
-      footer={{ height: 60 }}
+      header={{ height: "3rem" }}
       navbar={{ width: 300, breakpoint: "sm", collapsed: { mobile: !opened } }}
       aside={{
         width: 300,
@@ -41,17 +38,6 @@ const AppShell: React.FC<AppShellProps> = ({ children }) => {
         {children}
       </MantineAppShell.Main>
       <MantineAppShell.Aside p="md">Aside</MantineAppShell.Aside>
-      <MantineAppShell.Footer p="xs" h="auto" withBorder={false}>
-        <TextEditor
-          value="aaa"
-          onChange={() => {
-            return 0;
-          }}
-          onSend={() => {
-            return 0;
-          }}
-        />
-      </MantineAppShell.Footer>
     </MantineAppShell>
   );
 };
