+++ src/app/channels/[channel_id]/ChannelIDPage.module.css
@@ -1,0 +1,4 @@
+.root {
+  min-height: 100vh;
+  padding-top: 3rem;
+}
