+++ src/app/channels/[channel_id]/_components/ChannelFooter/ChannelFooter.module.css
@@ -1,0 +1,3 @@
+.channel-footer {
+  padding: 0.5rem;
+}
